remove unnecessary new-flow directory
