+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/ContributionThankYouHearMarketingConsent.jsx
@@ -19,7 +19,7 @@ import SvgNotification from './components/SvgNotification';
 import styles from './styles';
 import {
   OPHAN_COMPONENT_ID_MARKETING,
-} from '../utils/ophan';
+} from './utils/ophan';
 import { trackComponentClick } from 'helpers/tracking/behaviour';
 
 const checkboxContainer = css`
