+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/components/ExpandableContainer.jsx
