+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/components/SvgClock.jsx
